no pointer for datawrite

+++ spi_lib.h
@@ -76,8 +76,8 @@ typedef struct
 /* ***            Declaration of global functions                     ***/
 /* **********************************************************************/
 s32 	spi_init(u8* spiDevice, ST_SPI_IOC_SETTING_TYPE* st_spiIocSetting);
-ssize_t spi_write(s32 s32_spiFile, u8 *dataToWrite, u8 *numOfBytesToWrite);
-ssize_t spi_read(s32 s32_spiFile, u8 *dataToWrite, u8 *numOfBytesToWrite, u64* dataToRead, u8* numOfBytesToRead);
+ssize_t spi_write(s32 s32_spiFile, u64 *dataToWrite, u8 *numOfBytesToWrite);
+ssize_t spi_read(s32 s32_spiFile, u64 *dataToWrite, u8 *numOfBytesToWrite, u64* dataToRead, u8* numOfBytesToRead);
 void	spi_close(s32 s32_spiFile);
